Updated Category List with the ability to add a catergory removed route in application views
