+++ src/components/Categories/CategoriesList.js
@@ -9,6 +9,7 @@ import "./Categories.css"
 export const CategoryList = () => {
     const [categories, setCategories] = useState ([])
     // const [filteredPosts, setFiltered] = useState([])
+    const [value, setValue] = useState('')
     const navigate = useNavigate()
 
     // useEffect(
@@ -36,8 +37,40 @@ export const CategoryList = () => {
             .then((categoryArray) => {
             setCategories(categoryArray)
             })
-    }, [categories])
-
+    }, [])
+    const rerender = () => {
+        fetch(`http://localhost:8088/categories`)
+            .then((res) => res.json())
+            .then((categoryArray) => {
+            setCategories(categoryArray)
+            })
+    }
+   
+    const handleChange = (event) => {
+        setValue(event.target.value);
+        
+      };
+      const handleSaveButtonClick = (event) => {
+        event.preventDefault()
+        const categoryToSendToAPI = {
+            label: value
+            
+        }
+        return fetch(`http://localhost:8088/categories`, {
+            method: "POST",
+            headers: {
+                "Content-Type": "application/json"
+            },
+            body: JSON.stringify(categoryToSendToAPI)
+        })
+        .then (rerender)
+        
+  
+        
+        
+      
+    }
+    
         return (
             <>
             <div className="top-of-posts">
@@ -55,7 +88,13 @@ export const CategoryList = () => {
                         
                     )
                 })}
-                <AddCategory></AddCategory>
+                <form>
+        <label htmlFor="categoryadd">Add New Category</label>
+      <input className="categoryadd" type="text" value={value} onChange={handleChange} />
+      <button
+      onClick={(clickEvent) => handleSaveButtonClick(clickEvent)}
+      >Save Category</button>
+    </form>
                 </div>
             </>
         )
